1.3.0.1
It should now be possible to download the Vanilla pack again.

The launcher now downloads Minecrafts base files from Mojangs servers,
acording to their rules.
Better memory allocation settings.
The Pack selector is not a dropdown listm and its not possible to edit
the text in them. Silly noobish oversight.